requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
fake-useragent==1.5.1
Flask==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
outcome==1.3.0.post0
pillow==10.3.0
PyAutoGUI==0.9.54
pydub==0.25.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-Quartz==10.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytweening==1.2.0
pywhatkit==5.4
requests==2.31.0
rubicon-objc==0.4.9
selenium==4.8.3
selenium-recaptcha-solver==1.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==1.26.18
Werkzeug==3.0.3
wikipedia==1.4.0
wsproto==1.2.0
